Move coinbase smart wallet abi into its own abi file within _contracts.
