+++ template/web/app/paymaster-bundler/_contracts/coinbaseSmartWalletABI.ts
@@ -1,0 +1,32 @@
+export const coinbaseSmartWalletABI = [
+    {
+      type: 'function',
+      name: 'executeBatch',
+      inputs: [
+        {
+          name: 'calls',
+          type: 'tuple[]',
+          internalType: 'struct CoinbaseSmartWallet.Call[]',
+          components: [
+            {
+              name: 'target',
+              type: 'address',
+              internalType: 'address',
+            },
+            {
+              name: 'value',
+              type: 'uint256',
+              internalType: 'uint256',
+            },
+            {
+              name: 'data',
+              type: 'bytes',
+              internalType: 'bytes',
+            },
+          ],
+        },
+      ],
+      outputs: [],
+      stateMutability: 'payable',
+    },
+  ];
