+++ template/web/app/paymaster-bundler/constants.ts
@@ -9,36 +9,3 @@ export const coinbaseSmartWalletV1Implementation = '0x000100abaad02f1cfC8Bbe32bD
 export const magicSpendAddress = '0x011A61C07DbF256A68256B1cB51A5e246730aB92';
 export const erc1967ProxyImplementationSlot =
   '0x360894a13ba1a3210667c828492db98dca3e2076cc3735a920a3ca505d382bbc';
-
-export const coinbaseSmartWalletABI = [
-  {
-    type: 'function',
-    name: 'executeBatch',
-    inputs: [
-      {
-        name: 'calls',
-        type: 'tuple[]',
-        internalType: 'struct CoinbaseSmartWallet.Call[]',
-        components: [
-          {
-            name: 'target',
-            type: 'address',
-            internalType: 'address',
-          },
-          {
-            name: 'value',
-            type: 'uint256',
-            internalType: 'uint256',
-          },
-          {
-            name: 'data',
-            type: 'bytes',
-            internalType: 'bytes',
-          },
-        ],
-      },
-    ],
-    outputs: [],
-    stateMutability: 'payable',
-  },
-];
